+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 - Bug #73: Standardize PHPDoc headers across src directory files (@terabytesoftw)
 - Bug #74: Remove directory `tests\Stub` and move `tests\Support\Provider` to `tests\Provider` in tests (@terabytesoftw)
 - Bug #75: Standardize PHPDoc headers for test classes (@terabytesoftw)
+- Bug #76: Move `HasDisabled` trait to `UIAwesome\Html\Attribute` namespace and update related imports accordingly (@terabytesoftw)
 
 ## 0.5.2 January 29, 2026
