+++ tests/Provider/DisabledProvider.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace UIAwesome\Html\Attribute\Tests\Provider\Form;
+namespace UIAwesome\Html\Attribute\Tests\Provider;
 
 /**
  * Data provider for {@see \UIAwesome\Html\Attribute\Tests\Form\HasDisabledTest} test cases.
